refactored coral scoring/intaking command
